Renamed the "int" type to "i64".

+++ tests/invalid/conditional_statements/declare_statement.rs
@@ -2,7 +2,7 @@
 fail: 1
 */
 
-main(): int {
+main(): i32 {
     if (5)
-        i: int = 0;
+        i: i64 = 0;
 }
