+++ tests/invalid/conditional_statements/if_assignment.rs
@@ -2,9 +2,9 @@
 fail: 1
 */
 
-main(): int {
-    flag: int = 0;
-    a: int = if (flag)
+main(): i32 {
+    flag: i64 = 0;
+    a: i64 = if (flag)
                 2;
             else
                 3;
