+++ tests/invalid/functions/declaration_mismatch.rs
@@ -2,12 +2,12 @@
 fail: 1
 */
 
-foo(a: int): int;
+foo(a: i64): i64;
 
-main(): int {
+main(): i32 {
     return 5;
 }
 
-foo(a: int, b: int): int {
+foo(a: i64, b: i64): i64 {
     return 4;
 }
