+++ tests/invalid/functions/too_many_args.rs
@@ -2,10 +2,10 @@
 fail: 1
 */
 
-foo(a: int): int {
+foo(a: i64): i64 {
     return a + 1;
 }
 
-main(): int {
+main(): i32 {
     return foo(1, 2);
 }
