+++ tests/valid/conditional_expression/nested_ternary.rs
@@ -2,10 +2,10 @@
 return: 7
 */
 
-main(): int {
-    a: int = 1;
-    b: int = 2;
-    flag: int = 0;
+main(): i32 {
+    a: i64 = 1;
+    b: i64 = 2;
+    flag: i64 = 0;
 
     return a > b ? 5 : flag ? 6 : 7;
 }
