+++ tests/valid/conditional_statement/if_not_taken.rs
@@ -2,9 +2,9 @@
 return: 0
 */
 
-main(): int {
-    a: int = 0;
-    b: int = 0;
+main(): i32 {
+    a: i64 = 0;
+    b: i64 = 0;
     if (a)
         b = 1;
     return b;
