+++ tests/valid/functions/fib.rs
@@ -2,7 +2,7 @@
 return: 5
 */
 
-fib(n: int): int {
+fib(n: i64): i64 {
     if (n == 0 || n == 1) {
         return n;
     } else {
@@ -10,7 +10,7 @@ fib(n: int): int {
     }
 }
 
-main(): int {
-    n: int = 5;
+main(): i32 {
+    n: i64 = 5;
     return fib(n);
 }
