+++ tests/valid/functions/later_decl.rs
@@ -2,12 +2,12 @@
 return: 5
 */
 
-foo(a: int): int {
+foo(a: i64): i64 {
     return a + 1;
 }
 
-main(): int {
+main(): i32 {
     return foo(4);
 }
 
-foo(a: int): int;
+foo(a: i64): i64;
