+++ tests/valid/loops/nested_break.rs
@@ -2,10 +2,10 @@
 return: 250
 */
 
-main(): int {
-    ans: int = 0;
-    for (i: int = 0; i < 10; i = i + 1)
-        for (j: int = 0; j < 10; j = j + 1)
+main(): i32 {
+    ans: i64 = 0;
+    for (i: i64 = 0; i < 10; i = i + 1)
+        for (j: i64 = 0; j < 10; j = j + 1)
             if ((i / 2)*2 == i)
                 break;
             else
