+++ tests/valid/loops/while_multi_statement.rs
@@ -2,9 +2,9 @@
 return: 6
 */
 
-main(): int {
-    a: int = 0;
-    b: int = 1;
+main(): i32 {
+    a: i64 = 0;
+    b: i64 = 1;
 
     while (a < 5) {
         a = a + 2;
