+++ tests/valid/variables/assign_val.rs
@@ -2,8 +2,8 @@
 return: 0
 */
 
-main(): int {
-    a: int;
-    b: int = a = 0;
+main(): i32 {
+    a: i64;
+    b: i64 = a = 0;
     return b;
 }
